+++ src/Service/Lambda/src/Input/InvocationRequest.php
@@ -76,6 +76,13 @@ final class InvocationRequest extends Input
     private $qualifier;
 
     /**
+     * The identifier of the tenant in a multi-tenant Lambda function.
+     *
+     * @var string|null
+     */
+    private $tenantId;
+
+    /**
      * @param array{
      *   FunctionName?: string,
      *   InvocationType?: InvocationType::*|null,
@@ -83,6 +90,7 @@ final class InvocationRequest extends Input
      *   ClientContext?: string|null,
      *   Payload?: string|null,
      *   Qualifier?: string|null,
+     *   TenantId?: string|null,
      *   '@region'?: string|null,
      * } $input
      */
@@ -94,6 +102,7 @@ final class InvocationRequest extends Input
         $this->clientContext = $input['ClientContext'] ?? null;
         $this->payload = $input['Payload'] ?? null;
         $this->qualifier = $input['Qualifier'] ?? null;
+        $this->tenantId = $input['TenantId'] ?? null;
         parent::__construct($input);
     }
 
@@ -105,6 +114,7 @@ final class InvocationRequest extends Input
      *   ClientContext?: string|null,
      *   Payload?: string|null,
      *   Qualifier?: string|null,
+     *   TenantId?: string|null,
      *   '@region'?: string|null,
      * }|InvocationRequest $input
      */
@@ -149,6 +159,11 @@ final class InvocationRequest extends Input
         return $this->qualifier;
     }
 
+    public function getTenantId(): ?string
+    {
+        return $this->tenantId;
+    }
+
     /**
      * @internal
      */
@@ -173,6 +188,9 @@ final class InvocationRequest extends Input
         }
         if (null !== $this->clientContext) {
             $headers['X-Amz-Client-Context'] = $this->clientContext;
+        }
+        if (null !== $this->tenantId) {
+            $headers['X-Amz-Tenant-Id'] = $this->tenantId;
         }
 
         // Prepare query
@@ -240,6 +258,13 @@ final class InvocationRequest extends Input
     public function setQualifier(?string $value): self
     {
         $this->qualifier = $value;
+
+        return $this;
+    }
+
+    public function setTenantId(?string $value): self
+    {
+        $this->tenantId = $value;
 
         return $this;
     }