+++ src/Service/Lambda/CHANGELOG.md
@@ -8,6 +8,7 @@
 - AWS api-change: Added `us-isob-west-1` region
 - AWS api-change: Added SerializedRequestEntityTooLargeException to Lambda Invoke API
 - AWS api-change: Add Python3.14 (python3.14) and Java 25 (java25) support to AWS Lambda
+- AWS api-change: Added support for creating and invoking Tenant Isolated functions in AWS Lambda APIs.
 
 ### Dependency bumped